ovpn-dco: add PEER_ID to CMD_NEW_PEER and remove "REMOTE" attribute
Add PEER_ID field to CMD_NEW_PEER so that Peer ID can be decided in
userspace.

Also, peer IDs are stored in the peer object and are used for sending and
receiving traffic towards a specific peer. For this reason it makes more
sense to drop any reference to it being "remote" or not.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>