ovpn-dco: drop reference to socket if already owned
Each time a new peer is created, a reference to the socket passed as
argument is taken.

However, ovpn-dco assumes that only one reference is held per socket,
which is then dropped when all peers using that socket have been
deleted.

Due to the above, a reference unbalance is created as each new-peer call
takes a reference, but the release routine would drop only one.

Fix this unbalance by dropping the taken reference every time a peer is
created, but the passed socket was already owned by ovpn-dco

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>