ovpn-dco: in case of error bail out before adding peer to hashtables
When adding a peer it is expected that its bind member is already
initialized.

However, should this not be the case, the peer has to be removed from
any hashtable before bailing out.

Simplify the code the ensure the above condition is met by performing
the check on 'bind' before adding the peer to any table.

While at it, make sure that no reference counting is dropped for the
peer, as this context does not own any.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>