ovpn-dco: disable ICMP redirects by default
When a host is connected to multiple peers (i.e. an OpenVPN server) it
will be used as relay point between the various peers.

In this case, though, the kernel will believe that peers should rather
communicate with each other without passing through the server (because
it looks like they all share the same network/broadcast domain). For
this reason the kernel will send ICMP redirects back to the source peer.

Disable this behaviour by default as it is has no effect in the ovpn-dco
scenario and only creates annoying ICMP packets.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>